+++ src/app.js
@@ -1,6 +1,7 @@
 import React, { Component } from 'react'
 
 import Main from './containers/main'
+import Listing from './containers/listing'
 
 import Header from './components/header'
 import Footer from './components/footer'
@@ -10,7 +11,8 @@ export default class App extends Component {
     return (
     	<div>
 	    	<Header />
-	    	<Main />
+		    	<Main />
+		    	<Listing />
 	    	<Footer />
 	    </div>
     )