+++ src/containers/listing/index.js
@@ -1,0 +1,14 @@
+import React, { Component } from 'react';
+
+import styles from './styles.scss'
+
+export default class Listing extends Component {
+  
+  render() {
+    return (
+      <main id="listing" className="container">
+        <p>This is where the listings view will go</p>
+      </main>
+    );
+  }
+}